requirements.txt
celery==4.4.2
Django==4.1.3
django-celery-beat==2.4.0
newsapi-python==0.2.6
redis==4.3.5
